+++ packages/cli/create-damat-app/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @damatjs/create-damat-app
 
+## 0.4.0
+
+### Minor Changes
+
+- matching the version cleanup on the workflow package deps, and setup the migration cli for the modules
+
+### Patch Changes
+
+- Updated dependencies
+  - @damatjs/cli@0.4.0
+  - @damatjs/deps@0.4.0
+
 ## 0.3.7
 
 ### Patch Changes