+++ packages/core/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @damatjs/types
 
+## 0.4.0
+
+### Minor Changes
+
+- matching the version cleanup on the workflow package deps, and setup the migration cli for the modules
+
 ## 0.3.7
 
 ### Patch Changes