+++ packages/module/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @damatjs/module
 
+## 0.4.0
+
+### Minor Changes
+
+- matching the version cleanup on the workflow package deps, and setup the migration cli for the modules
+
+### Patch Changes
+
+- Updated dependencies
+  - @damatjs/workflow-engine@0.4.0
+  - @damatjs/logger@0.4.0
+  - @damatjs/deps@0.4.0
+  - @damatjs/framework@0.4.0
+  - @damatjs/orm-connector@0.4.0
+  - @damatjs/orm-migration@0.4.0
+  - @damatjs/orm-model@0.4.0
+  - @damatjs/orm-type@0.4.0
+  - @damatjs/services@0.4.0
+  - @damatjs/codegen@2.0.0
+
 ## 0.3.7
 
 ### Patch Changes