+++ packages/orm/connector/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @damatjs/orm-connector
 
+## 0.4.0
+
+### Minor Changes
+
+- matching the version cleanup on the workflow package deps, and setup the migration cli for the modules
+
+### Patch Changes
+
+- Updated dependencies
+  - @damatjs/logger@0.4.0
+  - @damatjs/types@0.4.0
+  - @damatjs/deps@0.4.0
+  - @damatjs/orm-type@0.4.0
+
 ## 0.3.7
 
 ### Patch Changes